extracted loginsignup form to a reusable component so we can keep our views DRY and ship a smaller bundle
