+++ client/src/components/LoginSignup/LoginSignup.scss
@@ -219,7 +219,8 @@ button:hover {
 
 .login100-form-title {
   //   font-family: Poppins-Medium;
-  font-size: 30px;
+  font-size: 3rem;
+  font-weight: 600;
   color: #fff;
   line-height: 1.2;
   text-align: center;
