+++ client/src/views/Login/Login.jsx
@@ -1,92 +1,24 @@
-import React, { useState } from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faUser, faLock } from '@fortawesome/free-solid-svg-icons';
+import React from 'react';
 
-import './Login.scss';
+// imports the dual purpose auth form
+import LoginSignup from '../../components/LoginSignup/LoginSignup';
 
 const Login = () => {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-
   /**
-   * this handleChange covers both the username and password
-   * it determines which one to update based on the event it receives
-   * if the event target has a type of password, then it updates password -- otherwise, it updates username
+   * NOTE TO SELF:
+   * PICK UP BY EXTRACTING ALL OF THE BELOW RETURN CODE INTO A SEPARATE COMPONENT
+   * THAT USERS TERNARY OPERATORS TO DETERMINE CLASSNAMES / BACKGROUNDS, ETC
+   * HANDLSUBMIT, USER/PASSWORD STATE, ETC WILL ALL LIVE IN THAT COMPONENT
+   *
+   * Login and Signup will then be separate views that do almost nothing besides prop drill a piece of state
+   * that prop will be props.currPage, and will let child component know whether to use login or signup in it's ternarys
+   * this will result in some classname changes, a dif page background, and a different handleSubmit (dif fetch req)
    */
-  const handleChange = (e) => {
-    if (e.target.type === 'password') {
-      setPassword(e.target.value);
-    } else {
-      setUsername(e.target.value);
-    }
-  };
-
-  const handleSubmit = () => {};
-
   return (
-    <div className="limiter">
-      <div className="container-login100">
-        <div className="wrap-login100">
-          <form className="login100-form validate-form">
-            <span className="login100-form-logo">
-              <img src="../../images/milestones_1.png" alt="apartments" />
-            </span>
-
-            <span className="login100-form-title p-b-34 p-t-27">Log in</span>
-
-            <div className="wrap-input100 validate-input" data-validate="Enter username">
-              <input
-                className="input100"
-                type="text"
-                name="username"
-                placeholder="Username"
-                value={username}
-                onChange={handleChange}
-              />
-              {/* these icons should be brought in for style appeal, but need to have the css modified so they float when the input is selected */}
-              <FontAwesomeIcon className="focus-input100" color="white" icon={faUser} />
-            </div>
-
-            <div className="wrap-input100 validate-input" data-validate="Enter password">
-              <input
-                className="input100"
-                type="password"
-                name="pass"
-                placeholder="Password"
-                value={password}
-                onChange={handleChange}
-              />
-              {/* these icons should be brought in for style appeal, but need to have the css modified so they float when the input is selected */}
-              <FontAwesomeIcon className="focus-input100" color="white" icon={faLock} />
-            </div>
-
-            <div className="contact100-form-checkbox">
-              <input
-                className="input-checkbox100"
-                id="ckb1"
-                type="checkbox"
-                name="remember-me"
-              />
-              <label className="label-checkbox100" htmlFor="ckb1">
-                Remember me
-              </label>
-            </div>
-
-            <div className="container-login100-form-btn">
-              <button type="submit" className="login100-form-btn">
-                Login
-              </button>
-            </div>
-
-            <div className="text-center p-t-3">
-              <a className="txt1" href="#">
-                Forgot Password?
-              </a>
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
+    <>
+      {/* tell the auth form to display the appropriate classes/text for a login view */}
+      <LoginSignup currView="login" />
+    </>
   );
 };
